refactor(slack): give the Slack markdown block helpers one home
Signals owned the markdown block builder and its character budget, and the
Slack app now needs both. Sharing them keeps the 12,000-character budget
stated once, where it cannot drift between the two products.
