+++ products/signals/backend/slack_formatting.py
@@ -4,11 +4,7 @@ import re
 from collections import Counter
 
 from posthog.dataclasses import frozen
-
-# A Slack `markdown` block takes Markdown directly, and Slack budgets 12,000 characters across
-# every markdown block in one message. A message carries one, and the headroom covers the blocks
-# around it.
-SLACK_MARKDOWN_TEXT_MAX_LEN = 11500
+from posthog.helpers.slack_markdown import SLACK_MARKDOWN_TEXT_MAX_LEN, truncate_slack_text
 
 # A summary places a chart inline with a markdown link targeting `chart:<chart_id>`. Slack cannot
 # place an image mid-sentence, and the link degrades badly if left alone: `chart:` is no scheme a
@@ -107,18 +103,6 @@ def is_safe_slack_http_url(value: object) -> bool:
     if not (value.startswith("http://") or value.startswith("https://")):
         return False
     return not any(char in value for char in ("<", ">", "|"))
-
-
-def truncate_slack_text(text: str, limit: int) -> str:
-    """Keep text below the block's character limit with headroom."""
-    if len(text) <= limit:
-        return text
-    return text[: limit - 1].rstrip() + "…"
-
-
-def slack_markdown_block(text: str) -> dict:
-    """The Slack block that renders Markdown, for text `prepare_slack_markdown` has already made safe."""
-    return {"type": "markdown", "text": text}
 
 
 def prepare_slack_markdown(text: str) -> str:
